vendor 2/django-celery-results-master/requirements/test.txt
case>=1.3.1
pytest>=3.0
pytest-django
pytz>dev
